+++ Runtime/Libraries/NMath.cs
@@ -6,14 +6,22 @@ namespace Nazio_LT.Tools.Core
     {
         #region Constants
 
+        /// <summary>Golden ratio.</summary>
         public const float GOLDEN_NUMB = 1.618033988749894f;
 
+        /// <summary>Euler's number.</summary>
+        public const float E = 2.71828182846f;
+
+        /// <summary> Value wich is near to zero and positive.</summary>
         public const float EPSILON = 0.03f;
+
+        /// <summary> Square value of Epsilon.</summary>
         public const float SQUARE_EPSILON = EPSILON * EPSILON;
 
         #endregion
 
-        public static bool AverageZero(float _value) => _value < EPSILON && _value > -EPSILON;
+        /// <summary>Return if the value is near to zero. (His absolute value is lower than epsilon)</summary>
+        public static bool AverageZero(float _value) => Mathf.Abs(_value) < EPSILON;
 
         #region Vector Bounding
 
@@ -69,29 +77,15 @@ namespace Nazio_LT.Tools.Core
 
         #endregion
 
-        public static Vector2 Vector2PerAxis(bool _firstX, float _value1, float _value2) => new Vector2(_firstX ? _value1 : _value2, _firstX ? _value2 : _value1);
-
         #region Angles
 
-        /// <summary>
-        /// Calulate the angle of a 2D vector in degree.
-        /// </summary>
-        /// <param name="_x"></param>
-        /// <param name="_y"></param>
-        /// <returns></returns>
+        /// <summary> Calculate the angle of a 2D vector in degree.</summary>
         public static float VectorAngleDEG(float _x, float _y) => Mathf.Rad2Deg * VectorAngleRAD(_x, _y);
 
-        /// <summary>
-        /// Calulate the angle of a 2D vector in radian.
-        /// </summary>
-        /// <param name="_x"></param>
-        /// <param name="_y"></param>
-        /// <returns></returns>
+        /// <summary> Calculate the angle of a 2D vector in radian.</summary>
         public static float VectorAngleRAD(float _x, float _y) => Mathf.Atan(_y / _x);
 
         #endregion
-
-        public static Color SetAlpha(Color _value, float _alpha) => new Color(_value.r, _value.g, _value.b, _alpha);
 
         public static float Remap(float _value, float _inputMin, float _inputMax, float _outputMin, float _outputMax) => Mathf.Lerp(_outputMin, _outputMax, Mathf.InverseLerp(_inputMin, _inputMax, _value));
     }