+++ Runtime/Models/Child.cs
@@ -4,6 +4,9 @@ namespace Nazio_LT.Tools.Core
     {
         private T parent;
 
+        /// <summary>
+        /// Initialise l'objet
+        /// </summary>
         public virtual void Init(T _parent) => parent = _parent;
 
         public T Parent => parent;