+++ Runtime/Models/ChildMonoBehaviour.cs
@@ -4,16 +4,16 @@ namespace Nazio_LT.Tools.Core
 {
     public abstract class ChildMonoBehaviour<ParentType> : MonoBehaviour
     {
-        private ParentType core;
+        private ParentType parent;
 
         /// <summary>
         /// Initialise l'objet
         /// </summary>
         public virtual void Init(ParentType _core)
         {
-            core = _core;
+            parent = _core;
         }
 
-        protected ParentType Core => core;
+        protected ParentType Parent => parent;
     }
 }