+++ Runtime/System/Audio/MusicManager.cs
@@ -11,8 +11,7 @@ namespace Nazio_LT.Tools.Core
         public int playlist;
     }
 
-    [RequireComponent(typeof(AudioSource))]
-    [AddComponentMenu("Nazio_LT/Core/MusicManager")]
+    [RequireComponent(typeof(AudioSource)), AddComponentMenu("Nazio_LT/Core/MusicManager")]
     public class MusicManager : Singleton<MusicManager>
     {
         [SerializeField] private NMusic[] music;